Refactor: component별 파일구조 변경
