--- src/component/presenter/buttons/EditBtn.jsx
@@ -1,17 +1,0 @@
-import styled from 'styled-components';
-
-const CustomButton = styled.button`
-  color: #95bfb4;
-  margin-right: 3rem;
-  border: 0;
-  outline: none;
-  cursor: pointer;
-  font-size: 2rem;
-  font-weight: bold;
-  background-color: transparent;
-  &:hover {
-    border-bottom: 1px solid #95bfb4;
-  }
-`;
-
-export default CustomButton;
