--- src/component/presenter/errors/AuthError.jsx
@@ -1,11 +1,0 @@
-import styled from 'styled-components';
-
-const AuthError = styled.div`
-  padding-top: 0.3rem;
-  height: 1.3rem;
-  width: 100%;
-  font-size: 1rem;
-  color: #ff4100;
-`;
-
-export default AuthError;
